refactor Employees page and layout
