--- src/pages/employeeProfile/EmployeeProfile.jsx
@@ -1,6 +1,0 @@
-import { useParams } from "react-router-dom";
-
-export const EmployeeProfile = () => {
-  const { employeeId } = useParams();
-  return <div>EmployeeProfile: {employeeId}</div>;
-};
