--- src/pages/employees/Employees.jsx
@@ -1,3 +1,0 @@
-export const Employees = () => {
-  return <div>Employees table</div>;
-};
